Add news parse interface and fix closing webClients
